# docs/swagger.yaml
basePath: /api
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  internal_handlers.CreateGroupRequest:
    properties:
      name:
        maxLength: 50
        minLength: 2
        type: string
    required:
    - name
    type: object
...
  internal_handlers.GroupResponse:
    properties:
      createdAt:
        type: string
      creator_id:
        type: integer
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      invite_code:
        type: string
      members:
        items:
          $ref: '#/definitions/vocalin-backend_internal_models.User'
        type: array
      name:
        type: string
      pinned_message:
        type: string
      pinned_message_author_id:
        type: integer
      timer_start_date:
        type: string
      timer_title:
        type: string
      updatedAt:
        type: string
    type: object
  internal_handlers.HomeGroupResponse:
    properties:
      createdAt:
        type: string
      creator_id:
        type: integer
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      invite_code:
        type: string
      members:
        items:
          $ref: '#/definitions/vocalin-backend_internal_models.User'
        type: array
      name:
        type: string
      pinned_message:
        type: string
      pinned_message_author_id:
        type: integer
      timer_start_date:
        type: string
      timer_title:
        type: string
      updatedAt:
        type: string
    type: object
  internal_handlers.HomeUserResponse:
    properties:
      avatar_url:
        type: string
      createdAt:
        type: string
      current_status:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      group_id:
        description: Nullable if not in a group
        type: integer
      id:
        type: integer
      nickname:
        type: string
      phone:
        type: string
      status_updated_at:
        type: string
      updatedAt:
        type: string
      wechat_id:
        type: string
    type: object
  internal_handlers.JoinGroupRequest:
    properties:
      invite_code:
        type: string
    required:
    - invite_code
    type: object
  internal_handlers.LoginRequest:
    properties:
      nickname:
        maxLength: 50
        minLength: 2
        type: string
      password:
        maxLength: 72
        minLength: 6
        type: string
    required:
    - nickname
    - password
    type: object
  internal_handlers.LoginResponse:
    properties:
      access_token:
        type: string
      access_token_expires_at:
        type: string
      refresh_token:
        type: string
      refresh_token_expires_at:
        type: string
      user:
        $ref: '#/definitions/vocalin-backend_internal_models.User'
    type: object
  internal_handlers.LogoutRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
...
  internal_handlers.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  internal_handlers.RefreshTokenResponse:
    properties:
      access_token:
        type: string
      access_token_expires_at:
        type: string
      refresh_token:
        type: string
      refresh_token_expires_at:
        type: string
    type: object
  internal_handlers.RegisterRequest:
    properties:
      confirm_password:
        maxLength: 72
        minLength: 6
        type: string
      nickname:
        maxLength: 50
        minLength: 2
        type: string
      password:
        maxLength: 72
        minLength: 6
        type: string
      phone:
        maxLength: 20
        minLength: 6
        type: string
    required:
    - confirm_password
    - nickname
    - password
    - phone
    type: object
  internal_handlers.UpdatePinnedMessageRequest:
    properties:
      content:
        maxLength: 500
        type: string
    required:
    - content
    type: object
  internal_handlers.UpdateStatusRequest:
    properties:
      status:
        maxLength: 120
        type: string
    required:
    - status
    type: object
  internal_handlers.UpdateTimerRequest:
    properties:
      start_date:
        type: string
      title:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - start_date
    - title
    type: object
...
  vocalin-backend_internal_models.Group:
    properties:
      createdAt:
        type: string
      creator_id:
        type: integer
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      invite_code:
        type: string
      members:
        items:
          $ref: '#/definitions/vocalin-backend_internal_models.User'
        type: array
      name:
        type: string
      pinned_message:
        type: string
      pinned_message_author_id:
        type: integer
      timer_start_date:
        type: string
      timer_title:
        type: string
      updatedAt:
        type: string
    type: object
...
  vocalin-backend_internal_models.User:
    properties:
      avatar_url:
        type: string
      createdAt:
        type: string
      current_status:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      group_id:
        description: Nullable if not in a group
        type: integer
      id:
        type: integer
      nickname:
        type: string
      phone:
        type: string
      status_updated_at:
        type: string
      updatedAt:
        type: string
      wechat_id:
        type: string
    type: object
  vocalin-backend_internal_response.APIResponse:
    properties:
      code:
        type: string
      data: {}
      message:
        type: string
      meta: {}
    type: object
...
  vocalin-backend_internal_service.DashboardResult:
    properties:
      group:
        $ref: '#/definitions/vocalin-backend_internal_models.Group'
      recent_activity: {}
    type: object
host: localhost:8080
info:
  contact: {}
  description: Vocalin（窝聚）后端服务，基于 Gin + GORM + Viper + Zap + Validator + JWT + Swagger
    构建
  title: Vocalin API
  version: "2.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: 当前使用昵称和密码登录，后续可扩展手机号验证码登录
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vocalin-backend_internal_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handlers.LoginResponse'
              type: object
      summary: 用户登录
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: 撤销当前 refresh token，使其无法再次刷新访问令牌
      parameters:
      - description: Logout Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vocalin-backend_internal_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 登出当前会话
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: 使用 refresh token 换取新的 access token 和 refresh token
      parameters:
      - description: Refresh Token Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vocalin-backend_internal_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handlers.RefreshTokenResponse'
              type: object
      summary: 刷新访问令牌
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: 使用昵称、手机号和密码创建账号，成功后直接返回 JWT
      parameters:
      - description: Register Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vocalin-backend_internal_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handlers.LoginResponse'
              type: object
      summary: 用户注册
      tags:
      - Auth
  /groups:
    post:
      consumes:
      - application/json
      description: 创建一个新的亲密空间，并自动将当前用户加入空间
      parameters:
      - description: Create Group Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vocalin-backend_internal_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handlers.GroupResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 创建空间
      tags:
      - Group
  /groups/create:
    post:
      consumes:
      - application/json
      description: 创建一个新的亲密空间，并自动将当前用户加入空间
      parameters:
      - description: Create Group Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vocalin-backend_internal_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handlers.GroupResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 创建空间
      tags:
      - Group
  /groups/join:
    post:
      consumes:
      - application/json
      description: 使用邀请码加入已有空间
      parameters:
      - description: Join Group Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.JoinGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vocalin-backend_internal_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handlers.GroupResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 加入空间
      tags:
      - Group
  /groups/me:
    get:
      description: 获取当前用户所在空间及成员信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vocalin-backend_internal_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handlers.GroupResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 获取当前空间信息
      tags:
      - Group
  /home/dashboard:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vocalin-backend_internal_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/vocalin-backend_internal_service.DashboardResult'
              type: object
      security:
      - BearerAuth: []
      summary: 获取首页概览
      tags:
      - Home
  /home/pinned:
    put:
      consumes:
      - application/json
      parameters:
      - description: Update Pinned Message Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.UpdatePinnedMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vocalin-backend_internal_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handlers.HomeGroupResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 更新置顶留言
      tags:
      - Home
  /home/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: Update Status Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.UpdateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vocalin-backend_internal_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handlers.HomeUserResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 更新实时状态
      tags:
      - Home
  /home/timer:
    put:
      consumes:
      - application/json
      parameters:
      - description: Update Timer Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.UpdateTimerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vocalin-backend_internal_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handlers.HomeGroupResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 更新陪伴计时器
      tags:
      - Home
...
securityDefinitions:
  BearerAuth:
    description: 使用 Bearer <token> 传递 JWT 访问令牌
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"